requirements.txt

autopep8==1.5.7
dj-database-url==0.5.0
gunicorn==20.1.0
pycodestyle==2.7.0
python-decouple==3.4
pytz==2021.1
sqlparse==0.4.2
toml==0.10.2
Unipath==1.1
whitenoise==5.3.0
asgiref==3.8.1
attrs==18.1.0
backports.zoneinfo==0.2.1
bcrypt==4.2.0
cacheout==0.11.0
certifi==2018.10.15
cffi==1.17.0
chardet==3.0.4
charset-normalizer==2.0.4
clickhouse-driver==0.2.3
cryptography==43.0.0
Django==4.2.15
django-commands==0.7
django-crontab==0.7.1
idna==2.7
influxdb==5.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
mongoengine==0.20.0
msgpack==1.0.2
packaging==21.3
paramiko==3.4.0
psutil==5.8.0
py-zabbix==1.1.7
pycparser==2.22
pymongo==3.7.2
PyNaCl==1.5.0
pyparsing==3.0.6
python-dateutil==2.7.5
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
requests==2.20.0
resolvelib==0.5.5
six==1.11.0
typing_extensions==4.12.2
tzdata==2022.1
tzlocal==4.1
urllib3==1.24.1

